Cleanup CVE-2007-4559 Patch
