+++ src/funfuzz/ccoverage/get_build.py
@@ -17,6 +17,7 @@ import fuzzfetch
 import requests
 
 from ..js.inspect_shell import queryBuildConfiguration
+from ..util.file_system_helpers import safe_tar_extractall
 
 RUN_COV_LOG = logging.getLogger("funfuzz")
 
@@ -97,29 +98,7 @@ def get_grcov(dirpath, args):
         grcov_bin_folder = dirpath / "grcov-bin"
         grcov_bin_folder.mkdir(parents=True, exist_ok=True)  # Ensure this dir has been created for Python 3.5 reasons
         with tarfile.open(fileobj=io.BytesIO(grcov_request.content), mode="r:bz2") as f:
-            
-            import os
-            
-            def is_within_directory(directory, target):
-                
-                abs_directory = os.path.abspath(directory)
-                abs_target = os.path.abspath(target)
-            
-                prefix = os.path.commonprefix([abs_directory, abs_target])
-                
-                return prefix == abs_directory
-            
-            def safe_extract(tar, path=".", members=None, *, numeric_owner=False):
-            
-                for member in tar.getmembers():
-                    member_path = os.path.join(path, member.name)
-                    if not is_within_directory(path, member_path):
-                        raise Exception("Attempted Path Traversal in Tar File")
-            
-                tar.extractall(path, members, numeric_owner=numeric_owner) 
-                
-            
-            safe_extract(f, str(grcov_bin_folder.resolve()))
+            safe_tar_extractall(f, str(grcov_bin_folder.resolve()))
 
     RUN_COV_LOG.info("grcov tarball extracted to this folder: %s", grcov_bin_folder.resolve())
     grcov_bin = grcov_bin_folder / f'grcov{".exe" if platform.system() == "Windows" else ""}'
